Support both current versions of resctl (2.1, 2.2)
Build and make both available as artefacts. This enables us to accept
results for either for the database and gives us a foundation for adding
new benchmark versions going forward.

This will also allow us to select at distribution time which set of
results to use when converting the internal database into configuration
to ship to distributions.