Refactor the code to split merging
Merging result files is something we should decouple from the issue
processing, as if we do it at that point there will be many conflicts
when people submit many new results for one model.

This will allow us to run a workflow specifically for merging, when the
changes are accepted.